add pre-0.3.0 changelog
I followed the example in `rasterio` (whose change logs I like) and labeled the changes with "next", to be updated to `0.3.0` once we merge everything we want to include and release. I also changed the extension since it is written with markdown already, and it seems like this improves the readability, but happy to change back if there is some reason I'm not seeing for it not having an extension.