ENH: add cx indexer to GeoDataFrame (#471) (#482)
* Add coordinate indexing method cx to geodataframes

Add a few basic tests that it works on a geodataframe as it does for a
geoseries.

* Add docs on indexing, including slicing with cx

* Use consistent assert in gdf coord slice test

* Move _CoordinateIndexer to base module

With both geoseries and geodataframe modules using the class, base.py is
a better place to define the _CoordinateIndexer class.

Also add docs to the _CoordinateIndexer class.

(cherry picked from commit 55d2d47d534a9bbbe7614c8faf9b3600475ed7f7)