ENH: Set point markersize with column values
Can optionally set the size of points in a scatter plot of point
geometries to be specified by the value of a column in the frame. Note
that it does not perform any normalization at present, so very small or
very large values in the column specified lead to less than desirable
results. Also no checks on user, so if a user specifies a column with a
non numeric type, it does throw a slightly less than clear error.