CLN: Use read_file from geodataframe in geoseries
I started looking at the issue raised in #881, and realized there was a
somewehat extraneous codepath with `fiona.open` in the geoseries
`read_file` method. This gives one fewer place to make
updates to use `fiona.Env()`.

It seems a little more natural to me to use the
`io.file.read_file` method here, but the adjoining two class methods are
importing from geodataframe and using the methods on `GeoDataFrame`
objects, so I followed suit here.