Set geocodefarm as default geocoding service
I hesitated to do this before for the docs, but it is nice to have the
default provider be one that works without an API key (which is now
required for google). The downside of this is that geocodefarm has only
been an option since May 2018, so if someone tries just running the
examples and doesn't have prior experience, they'll get a crash due
to a missing provider rather than a crash due a missing API key.