Remove unused make_valid argument
While trying to figure out an issue with an overlay returning surprising
results, I realized that the make_valid argument wasn't actually used
anywhere. I can imagine some cases where there would be some utility in
that, but let's remove it in the meantime. A question though: do we need
to provide a deprecation warning first for removing an argument that
hasn't actually been doing anything (and a quick search through history
makes it unclear whether it has ever done anything, but I'm not sure
about that).