Don't overwrite user-supplied geometry dtype
When writing to PostGIS, I want the option to allow mixed geometries in a table. The current behaviour of `to_postgis` implicitly disallows this by setting the geometry type to whatever the first dataframe type is to populate that table.

This change simply looks to see if the user has defined the geometry dtype first. If not then it continues with the original behaviour of setting the geometry type to the dataframe, otherwise it leaves the dtype untouched and allows the user to specify the target table geometry type.